add ducks pattern in project
