+++ src/index.js
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from 'react';
 import ReactDOM from 'react-dom/client';
-import * as actions from "./store/actions";
-import { initiateStore } from "./store/store";
+import configureStore from "./store/store";
+import { taskCompleted, taskDeleted, titleChanged } from "./store/task";
 
-const store = initiateStore();
+const store = configureStore();
 
 const App = () => {
     const [state, setState] = useState( store.getState() );
@@ -15,13 +15,13 @@ const App = () => {
     }, [] );
 
     const completeTask = ( taskId ) => {
-        store.dispatch( actions.taskCompleted( taskId ) );
+        store.dispatch( taskCompleted( taskId ) );
     };
     const changeTitle = ( taskId ) => {
-        store.dispatch( actions.titleChanged( taskId ) );
+        store.dispatch( titleChanged( taskId ) );
     }
     const deleteTask = ( taskId ) => {
-        store.dispatch( actions.taskDeleted( taskId ) );
+        store.dispatch( taskDeleted( taskId ) );
     };
 
     return (
