+++ src/store/store.js
@@ -1,11 +1,13 @@
 import { createStore } from "redux";
-import { taskReducer } from "./taskReducer";
+import taskReducer from "./task";
 
 const initialState = [
     {id: 1, title: "Task 1", completed: false},
     {id: 2, title: "Task 2", completed: false}
 ];
 
-export function initiateStore() {
+function configureStore() {
     return createStore(taskReducer, initialState);
 }
+
+export default configureStore;
